fix bugs backend

+++ backend/src/middlewares/deleteOldImageHandler.ts
@@ -1,22 +1,22 @@
-import { Request, Response, NextFunction } from "express";
+import { Request, Response, NextFunction, RequestHandler } from "express";
 import { unlink } from "fs/promises";
 import { getProduct } from "../services/productService";
 import { CustomError } from "../models/CustomError";
 
-export const deleteOldImageHandler = async (
+const deleteOldImageHandler = async (
   req: Request,
   res: Response,
   next: NextFunction
 ) => {
-  const uploadDir = process.env.UPLOADS_DIR;
-  const publicPath = process.env.PUBLIC_PATH;
+  const uploadDir = process.env.UPLOADS_DIR as string;
+  const publicPath = process.env.PUBLIC_PATH as string;
 
   // If there is a new image file, the old file is removed from the server.
   if (req.file) {
     try {
       const product = await getProduct(req.params.id);
       // If the old file is 'no_image.jpg', we do not delete it.
-      if (!product.imagePath.includes("no_image.jpg")) {
+      if (product && !product.imagePath.includes("no_image.jpg")) {
         await unlink(product.imagePath.replace("static", "public"));
       }
       req.body.image = req.file.path
@@ -29,3 +29,5 @@ export const deleteOldImageHandler = async (
 
   next();
 };
+
+export default deleteOldImageHandler as RequestHandler;
